+++ packages/component-library/src/components/feedback-indicator/mt-progress-bar/mt-progress-bar.vue
@@ -126,8 +126,6 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import "../../assets/scss/variables.scss";
-
 .mt-progress-bar {
   .mt-block-field__block {
     border: none;