+++ packages/component-library/src/components/form/mt-button/mt-button.vue
@@ -145,8 +145,6 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import "../../assets/scss/variables.scss";
-
 $mt-button-transition: all 0.15s ease-out;
 
 .mt-button {