+++ packages/component-library/src/components/table-and-list/mt-data-table/mt-data-table.stories.ts
@@ -315,6 +315,7 @@ export default {
     ],
     disableEdit: false,
     disableDelete: false,
+    disableSettingsTable: false,
     filters: [],
     appliedFilters: [],
     numberOfResults: undefined,