+++ packages/component-library/src/components/table-and-list/mt-data-table/sub-components/mt-data-table-filter/mt-data-table-filter.vue
@@ -90,8 +90,6 @@ export default defineComponent({
 </script>
 
 <style lang="scss" scoped>
-@import "../../../../assets/scss/variables.scss";
-
 .mt-data-table-filter {
   display: inline-flex;
   font-size: 12px;